issue 3: handles the % character create helper functions
